DataProcess/requirements_vertex.txt
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
tqdm>=4.66.0 